+++ sentinel/event_stream.go
@@ -8,12 +8,15 @@ import (
 	"strings"
 	"syscall"
 
+	"cosmossdk.io/errors"
 	"github.com/gogo/protobuf/proto"
 
 	"github.com/arkeonetwork/arkeo/common"
+	"github.com/arkeonetwork/arkeo/common/cosmos"
 
 	"github.com/cometbft/cometbft/libs/log"
 
+	tmlog "github.com/cometbft/cometbft/libs/log"
 	tmclient "github.com/cometbft/cometbft/rpc/client/http"
 	tmCoreTypes "github.com/cometbft/cometbft/rpc/core/types"
 	tmtypes "github.com/cometbft/cometbft/types"
@@ -21,6 +24,8 @@ import (
 
 	"github.com/arkeonetwork/arkeo/x/arkeo/types"
 )
+
+var numOfWebSocketClients = 2
 
 func subscribe(client *tmclient.HTTP, logger log.Logger, query string) <-chan tmCoreTypes.ResultEvent {
 	out, err := client.Subscribe(context.Background(), "", query)
@@ -31,26 +36,46 @@ func subscribe(client *tmclient.HTTP, logger log.Logger, query string) <-chan tm
 	return out
 }
 
+func NewTendermintClient(baseURL string) (*tmclient.HTTP, error) {
+	client, err := tmclient.New(baseURL, "/websocket")
+	if err != nil {
+		return nil, errors.Wrapf(err, "error creating websocket client")
+	}
+	logger := tmlog.NewTMLogger(tmlog.NewSyncWriter(os.Stdout))
+	client.SetLogger(logger)
+
+	return client, nil
+}
+
 func (p Proxy) EventListener(host string) {
 	logger := p.logger
-	client, err := tmclient.New(fmt.Sprintf("tcp://%s", host), "/websocket")
-	if err != nil {
-		logger.Error("failure to create websocket client", "error", err)
-		panic(err)
-	}
-	client.SetLogger(logger)
-	err = client.Start()
-	if err != nil {
-		logger.Error("Failed to start a client", "err", err)
-		os.Exit(1)
-	}
-	defer client.Stop() // nolint
 
-	// create a unified channel for receiving events
+	logger.Info("starting realtime indexing using /websocket")
 
+	// as maximum allowed connection is 5 per ws client(cometbft) we split this into 2 client to handle 3 connection each
+	clients := make([]*tmclient.HTTP, numOfWebSocketClients)
+
+	for i := 0; i < numOfWebSocketClients; i++ {
+		client, err := NewTendermintClient(fmt.Sprintf("tcp://%s", host))
+		if err != nil {
+			panic(fmt.Sprintf("error creating tm client for %s: %+v", host, err))
+		}
+		if err = client.Start(); err != nil {
+			panic(fmt.Sprintf("error starting ws client: %s: %+v", host, err))
+		}
+		defer func() {
+			if err := client.Stop(); err != nil {
+				logger.Error("Failed to stop the client", "error", err)
+			}
+		}()
+		clients[i] = client
+	}
+
+	// Create a unified channel for receiving events
 	eventChan := make(chan tmCoreTypes.ResultEvent, 1000)
 
-	subscribeToEvents := func(queries ...string) {
+	// Function to subscribe to events for a given client
+	subscribeToEvents := func(client *tmclient.HTTP, queries ...string) {
 		for _, query := range queries {
 			out := subscribe(client, logger, query)
 
@@ -59,7 +84,6 @@ func (p Proxy) EventListener(host string) {
 					select {
 					case result := <-out:
 						eventChan <- result
-
 					case <-client.Quit():
 						return
 					}
@@ -68,12 +92,17 @@ func (p Proxy) EventListener(host string) {
 		}
 	}
 
-	// subscribe to events
-	go subscribeToEvents(
+	// Subscribe to events for each client
+	go subscribeToEvents(clients[0],
 		"tm.event = 'NewBlock'",
 		"tm.event = 'Tx' AND message.action='/arkeo.arkeo.MsgOpenContract'",
 		"tm.event = 'Tx' AND message.action='/arkeo.arkeo.MsgCloseContract'",
+	)
+
+	go subscribeToEvents(clients[1],
 		"tm.event = 'Tx' AND message.action='/arkeo.arkeo.MsgClaimContractIncome'",
+		"tm.event = 'Tx' AND message.action='/arkeo.arkeo.MsgBondProvider'",
+		"tm.event = 'Tx' AND message.action='/arkeo.arkeo.MsgModProvider'",
 	)
 
 	dispatchEvents := func(result tmCoreTypes.ResultEvent) {
@@ -89,6 +118,12 @@ func (p Proxy) EventListener(host string) {
 
 		case strings.Contains(result.Query, "MsgClaimContractIncome"):
 			p.handleContractSettlementEvent(result)
+
+		case strings.Contains(result.Query, "MsgModProvider"):
+			p.handleModProviderEvent(result)
+
+		case strings.Contains(result.Query, "MsgBondProvider"):
+			p.handleBondProviderEvent(result)
 
 		default:
 			logger.Error("Unknown Event Type", "Query", result.Query)
@@ -280,4 +315,93 @@ func parseTypedEvent(result tmCoreTypes.ResultEvent, eventType string) (proto.Me
 	}
 
 	return msg, fmt.Errorf("event %s not found", eventType)
+}
+
+func (p Proxy) handleBondProviderEvent(result tmCoreTypes.ResultEvent) {
+	typedEvent, err := parseTypedEvent(result, "arkeo.arkeo.EventBondProvider")
+	if err != nil {
+		p.logger.Error("failed to parse typed event", "error", err)
+		return
+	}
+
+	evt, ok := typedEvent.(*types.EventBondProvider)
+	if !ok {
+		p.logger.Error(fmt.Sprintf("failed to cast %T to EventOpenContract", typedEvent))
+		return
+	}
+
+	service := common.Service(common.ServiceLookup[evt.Service])
+	if !p.isMyPubKey(evt.Provider) {
+		return
+	}
+	providerConfig, err := p.ProviderConfigStore.Get(evt.Provider, service.String())
+	if err != nil {
+		p.logger.Info("failed to get provider config, initializing new config", "error", err)
+		providerConfig = ProviderConfiguration{
+			PubKey:              evt.Provider,
+			Service:             service,
+			Bond:                evt.BondAbs,
+			BondRelative:        evt.BondRel,
+			MetadataUri:         "",
+			MetadataNonce:       0,
+			Status:              types.ProviderStatus(0),
+			MinContractDuration: 0,
+			MaxContractDuration: 0,
+			SubscriptionRate:    cosmos.Coins{},
+			PayAsYouGoRate:      cosmos.Coins{},
+			SettlementDuration:  0,
+		}
+	}
+
+	providerConfig.Bond = evt.BondAbs
+	providerConfig.BondRelative = evt.BondRel
+	err = p.ProviderConfigStore.Set(providerConfig)
+	if err != nil {
+		p.logger.Error("failed to update provider configuration", "error", err)
+		return
+	}
+	p.logger.Info("Provider configuration updated on bond provider event", "pubkey", evt.Provider.String(), "service", service.String())
+}
+func (p Proxy) handleModProviderEvent(result tmCoreTypes.ResultEvent) {
+	typedEvent, err := parseTypedEvent(result, "arkeo.arkeo.EventModProvider")
+	if err != nil {
+		p.logger.Error("failed to parse typed event", "error", err)
+		return
+	}
+
+	evt, ok := typedEvent.(*types.EventModProvider)
+	if !ok {
+		p.logger.Error(fmt.Sprintf("failed to cast %T to EventOpenContract", typedEvent))
+		return
+	}
+
+	service := common.Service(common.ServiceLookup[evt.Service])
+
+	if !p.isMyPubKey(evt.Provider) {
+		return
+	}
+
+	providerConfig, err := p.ProviderConfigStore.Get(evt.Provider, service.String())
+	if err != nil {
+		p.logger.Error(fmt.Sprintf("failed to get provider %s", err))
+		return
+	}
+
+	providerConfig.Bond = evt.Bond
+	providerConfig.Service = service
+	providerConfig.MetadataUri = evt.MetadataUri
+	providerConfig.MetadataNonce = evt.MetadataNonce
+	providerConfig.Status = evt.Status
+	providerConfig.MinContractDuration = evt.MinContractDuration
+	providerConfig.MaxContractDuration = evt.MaxContractDuration
+	providerConfig.SubscriptionRate = evt.SubscriptionRate
+	providerConfig.PayAsYouGoRate = evt.PayAsYouGoRate
+	providerConfig.SettlementDuration = evt.SettlementDuration
+
+	err = p.ProviderConfigStore.Set(providerConfig)
+	if err != nil {
+		p.logger.Error("failed to update provider configuration", "error", err)
+		return
+	}
+	p.logger.Info("Provider configuration updated on mod provider event", "pubkey", evt.Provider.String(), "service", service.String())
 }