+++ sentinel/routes.go
@@ -6,4 +6,5 @@ const (
 	RoutesClaim          = "/claim/{id}"
 	RoutesOpenClaims     = "/open-claims"
 	RouteManage          = "/manage/contract/{id}"
+	RouteProviderData    = "/provider/{service}"
 )