+++ CHANGELOG.md
@@ -28,6 +28,18 @@ Contains bug fixes.
 Contains all the PRs that improved the code without changing the behaviors.
 -->
 
+# v1.0.3-Prerelease 
+
+## Added 
+- Added sentinel setup docs
+
+## Changed
+- Updated sentinel to handle provider events
+
+## Fixed 
+- Fixed code lint
+- Fixed ws client issue with event stream
+
 # v1.0.2-Prerelease 
 
 ## Added